% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_footnote_in_fmt}
\alias{add_footnote_in_fmt}
\title{Add footnote in format}
\usage{
add_footnote_in_fmt(
  table_xx,
  footnotes_chr,
  output_type_1L_chr = c("PDF", "HTML", "Word"),
  footnote_fns_ls = make_table_fns_ls("null")
)
}
\arguments{
\item{table_xx}{Table (an output object of multiple potential types)}

\item{footnotes_chr}{Footnotes (a character vector)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: c("PDF", "HTML", "Word")}

\item{footnote_fns_ls}{Footnote functions (a list), Default: make_table_fns_ls("null")}
}
\value{
Table (an output object of multiple potential types)
}
\description{
add_footnote_in_fmt() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add footnote in format. The function returns Table (an output object of multiple potential types).
}
\keyword{internal}
