% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author-Ready4showSynopsis}
\alias{author-Ready4showSynopsis}
\alias{author,Ready4showSynopsis-method}
\title{Author and save files}
\usage{
\S4method{author}{Ready4showSynopsis}(x, results_ls = NULL)
}
\arguments{
\item{x}{An object of class Ready4showSynopsis}

\item{results_ls}{Results (a list), Default: NULL}
}
\description{
author method applied to Ready4showSynopsis
}
