% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData-Ready4showSynopsis}
\alias{authorData-Ready4showSynopsis}
\alias{authorData,Ready4showSynopsis-method}
\title{Author and document datasets}
\usage{
\S4method{authorData}{Ready4showSynopsis}(x)
}
\arguments{
\item{x}{An object of class Ready4showSynopsis}
}
\description{
authorData method applied to Ready4showSynopsis
}
