% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorReport.R
\name{authorReport-Ready4showSynopsis}
\alias{authorReport-Ready4showSynopsis}
\alias{authorReport,Ready4showSynopsis-method}
\title{Author and save a report}
\usage{
\S4method{authorReport}{Ready4showSynopsis}(x)
}
\arguments{
\item{x}{An object of class Ready4showSynopsis}
}
\description{
authorReport method applied to Ready4showSynopsis
}
