% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_authors_tb.R
\docType{data}
\name{authors_tb}
\alias{authors_tb}
\title{Example authors table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 10 columns.
}
\source{
\url{https://ready4-dev.com}
}
\usage{
authors_tb
}
\description{
Example of an authors table with fake author entries
}
\details{
A tibble

\describe{
\item{first_nm_chr}{First name (a character vector)}
\item{middle_nm_chr}{Middle name (a character vector)}
\item{last_nm_chr}{Last name (a character vector)}
\item{title_chr}{Title (a character vector)}
\item{qualifications_chr}{Qualifications (a character vector)}
\item{institute_chr}{Institute (a character vector)}
\item{sequence_int}{Sequence (an integer vector)}
\item{is_corresponding_lgl}{Is corresponding (a logical vector)}
\item{email_chr}{Email (a character vector)}
\item{is_equal_first_lgl}{Is equal first (a logical vector)}
}
}
\keyword{datasets}
