% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_institutes_tb.R
\docType{data}
\name{institutes_tb}
\alias{institutes_tb}
\title{Example institutes table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.
}
\source{
\url{https://ready4-dev.com}
}
\usage{
institutes_tb
}
\description{
Example of an institutes table with fake institute entries
}
\details{
A tibble

\describe{
\item{short_name_chr}{Short name (a character vector)}
\item{long_name_chr}{Long name (a character vector)}
}
}
\keyword{datasets}
