% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_authors_lup.R
\name{is_ready4_authors_lup}
\alias{is_ready4_authors_lup}
\title{Is ready4show S3 class for authors lookup table}
\usage{
is_ready4_authors_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4show S3 class for authors lookup table
}
\description{
Check whether an object is a valid instance of the ready4show S3 class for authors lookup table
}
\details{
ready4show S3 class for authors lookup table
}
