% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_institutes.R
\name{is_ready4show_institutes}
\alias{is_ready4show_institutes}
\title{is ready4show institutes ready4 S3 class for institutes lookup table}
\usage{
is_ready4show_institutes(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for institutes lookup table
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for institutes lookup table
}
\details{
ready4 S3 class for institutes lookup table
}
