% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_abstract_args_ls}
\alias{make_abstract_args_ls}
\title{Make abstract arguments list}
\usage{
make_abstract_args_ls(
  background_1L_chr = NA_character_,
  objectives_1L_chr = NA_character_,
  methods_1L_chr = NA_character_,
  results_1L_chr = NA_character_,
  conclusions_1L_chr = NA_character_,
  data_1L_chr = NA_character_,
  fl_nm_1L_chr = "abstract.txt"
)
}
\arguments{
\item{background_1L_chr}{Background (a character vector of length one), Default: 'NA'}

\item{objectives_1L_chr}{Objectives (a character vector of length one), Default: 'NA'}

\item{methods_1L_chr}{Methods (a character vector of length one), Default: 'NA'}

\item{results_1L_chr}{Results (a character vector of length one), Default: 'NA'}

\item{conclusions_1L_chr}{Conclusions (a character vector of length one), Default: 'NA'}

\item{data_1L_chr}{Data (a character vector of length one), Default: 'NA'}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'abstract.txt'}
}
\value{
Abstract arguments (a list)
}
\description{
make_abstract_args_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make abstract arguments list. The function returns Abstract arguments (a list).
}
\keyword{internal}
