% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_abstract_lines}
\alias{make_abstract_lines}
\title{Make abstract lines}
\usage{
make_abstract_lines(abstract_ls)
}
\arguments{
\item{abstract_ls}{Abstract (a list)}
}
\value{
Abstract lines (a character vector)
}
\description{
make_abstract_lines() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make abstract lines. The function returns Abstract lines (a character vector).
}
\keyword{internal}
