% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_authorship_lines}
\alias{make_authorship_lines}
\title{Make authorship lines}
\usage{
make_authorship_lines(authors_tb, inc_quals_1L_lgl = F)
}
\arguments{
\item{authors_tb}{Authors (a tibble)}

\item{inc_quals_1L_lgl}{Include qualifications (a logical vector of length one), Default: F}
}
\value{
Authorship lines (a character vector)
}
\description{
make_authorship_lines() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make authorship lines. The function returns Authorship lines (a character vector).
}
\keyword{internal}
