% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_default_paths}
\alias{make_default_paths}
\title{Make default paths}
\usage{
make_default_paths()
}
\value{
X (Metadata about paths to Markdown input and output)
}
\description{
make_default_paths() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make default paths. The function is called for its side effects and does not return a value.
}
\keyword{internal}
