% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_eq_ref}
\alias{make_eq_ref}
\title{make equation reference}
\usage{
make_eq_ref(eq_nm_1L_chr, output_type_1L_chr)
}
\arguments{
\item{eq_nm_1L_chr}{Equation name (a character vector of length one)}

\item{output_type_1L_chr}{Output type (a character vector of length one)}
}
\value{
Equation reference (a character vector of length one)
}
\description{
make_eq_ref() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make equation reference. The function returns Equation reference (a character vector of length one).
}
