% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_eqn_ref}
\alias{make_eqn_ref}
\title{Make eqn reference}
\usage{
make_eqn_ref(eqn_nm_1L_chr, output_type_1L_chr)
}
\arguments{
\item{eqn_nm_1L_chr}{Eqn name (a character vector of length one)}

\item{output_type_1L_chr}{Output type (a character vector of length one)}
}
\value{
Eqn reference (a character vector of length one)
}
\description{
make_eqn_ref() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make eqn reference. The function returns Eqn reference (a character vector of length one).
}
