% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_institutes_lines}
\alias{make_institutes_lines}
\title{Make institutes lines}
\usage{
make_institutes_lines(authors_tb, institutes_tb)
}
\arguments{
\item{authors_tb}{Authors (a tibble)}

\item{institutes_tb}{Institutes (a tibble)}
}
\value{
Institutes lines (a character vector)
}
\description{
make_institutes_lines() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make institutes lines. The function returns Institutes lines (a character vector).
}
\keyword{internal}
