% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_authors_lup.R
\name{make_new_ready4_authors_lup}
\alias{make_new_ready4_authors_lup}
\title{Make new ready4show S3 class for authors lookup table}
\usage{
make_new_ready4_authors_lup(x)
}
\arguments{
\item{x}{A prototype for the ready4show S3 class for authors lookup table}
}
\value{
An unvalidated instance of the ready4show S3 class for authors lookup table
}
\description{
Create a new unvalidated instance of the ready4show S3 class for authors lookup table
}
\details{
ready4show S3 class for authors lookup table
}
