% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_authors.R
\name{make_new_ready4show_authors}
\alias{make_new_ready4show_authors}
\title{make new ready4show authors ready4 submodule class for authors lookup table}
\usage{
make_new_ready4show_authors(x)
}
\arguments{
\item{x}{A prototype for the ready4 submodule class for authors lookup table}
}
\value{
An unvalidated instance of the ready4 submodule class for authors lookup table
}
\description{
Create a new unvalidated instance of the ready4 submodule class for authors lookup table
}
\details{
Authors lookup table
}
\keyword{internal}
