% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_correspondences.R
\name{make_new_ready4show_correspondences}
\alias{make_new_ready4show_correspondences}
\title{make new ready4show correspondences Name correspondences lookup table}
\usage{
make_new_ready4show_correspondences(x)
}
\arguments{
\item{x}{A prototype for the Name correspondences lookup table}
}
\value{
An unvalidated instance of the Name correspondences lookup table
}
\description{
Create a new unvalidated instance of the Name correspondences lookup table
}
\details{
Name correspondences lookup table
}
\keyword{internal}
