% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_institutes_lup.R
\name{make_new_ready4show_institutes_lup}
\alias{make_new_ready4show_institutes_lup}
\title{Make new ready4show institutes lookup table ready4 S3 class for institutes lookup table}
\usage{
make_new_ready4show_institutes_lup(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for institutes lookup table}
}
\value{
An unvalidated instance of the ready4 S3 class for institutes lookup table
}
\description{
Create a new unvalidated instance of the ready4 S3 class for institutes lookup table
}
\details{
ready4 S3 class for institutes lookup table
}
