% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_institutes.R
\name{make_pt_ready4show_institutes}
\alias{make_pt_ready4show_institutes}
\title{Make prototype ready4show institutes ready4 S3 class for institutes lookup table}
\usage{
make_pt_ready4show_institutes(
  short_name_chr = character(0),
  long_name_chr = character(0)
)
}
\arguments{
\item{short_name_chr}{Short name (a character vector), Default: character(0)}

\item{long_name_chr}{Long name (a character vector), Default: character(0)}
}
\value{
A prototype for ready4 S3 class for institutes lookup table
}
\description{
Create a new prototype for the ready4 S3 class for institutes lookup table
}
\details{
ready4 S3 class for institutes lookup table
}
