% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_rmd_fl_nms_ls}
\alias{make_rmd_fl_nms_ls}
\title{Make Markdown file names list}
\usage{
make_rmd_fl_nms_ls(
  html_fl_nm_1L_chr = "Main_HTML",
  pdf_fl_nm_1L_chr = "Main_Bookdown_PDF",
  word_fl_nm_1L_chr = "Main_Word"
)
}
\arguments{
\item{html_fl_nm_1L_chr}{Html file name (a character vector of length one), Default: 'Main_HTML'}

\item{pdf_fl_nm_1L_chr}{Pdf file name (a character vector of length one), Default: 'Main_Bookdown_PDF'}

\item{word_fl_nm_1L_chr}{Word file name (a character vector of length one), Default: 'Main_Word'}
}
\value{
a R Markdown file names (a list)
}
\description{
make_rmd_fl_nms_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make markdown file names list. The function returns a R Markdown file names (a list).
}
\keyword{internal}
