% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_rprt_type_ls}
\alias{make_rprt_type_ls}
\title{Make report type}
\usage{
make_rprt_type_ls(rprt_nm_1L_chr, rprt_lup)
}
\arguments{
\item{rprt_nm_1L_chr}{Report name (a character vector of length one)}

\item{rprt_lup}{Report (a lookup table)}
}
\value{
Report type (a list)
}
\description{
make_rprt_type_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make report type list. The function returns Report type (a list).
}
\keyword{internal}
