% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sngl_author_lines}
\alias{make_sngl_author_lines}
\title{Make single author lines}
\usage{
make_sngl_author_lines(authors_tb, slice_1L_int, inc_quals_1L_lgl = F)
}
\arguments{
\item{authors_tb}{Authors (a tibble)}

\item{slice_1L_int}{Slice (an integer vector of length one)}

\item{inc_quals_1L_lgl}{Include quals (a logical vector of length one), Default: F}
}
\value{
Author lines (a character vector)
}
\description{
make_sngl_author_lines() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make single author lines. The function returns Author lines (a character vector).
}
\keyword{internal}
