% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_from_chunk}
\alias{print_from_chunk}
\title{Print from chunk}
\usage{
print_from_chunk(
  ds_tb,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T,
  var_desc_chr = NA_character_,
  ...
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: NULL}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{use_lbls_as_col_nms_1L_lgl}{Use labels as column names (a logical vector of length one), Default: T}

\item{var_desc_chr}{Variable description (a character vector), Default: 'NA'}

\item{...}{Additional arguments}
}
\description{
print_from_chunk() is a Print function that prints output to console Specifically, this function implements an algorithm to print from chunk. The function is called for its side effects and does not return a value.
}
\keyword{internal}
