% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_in_format}
\alias{print_in_format}
\title{Print in format}
\usage{
print_in_format(
  table_xx,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = c("PDF", "HTML", "Word"),
  table_fns_ls = make_table_fns_ls(),
  use_lbls_as_col_nms_1L_lgl = T,
  var_desc_chr = NA_character_,
  ...
)
}
\arguments{
\item{table_xx}{Table (an output object of multiple potential types)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: NULL}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: c("PDF", "HTML", "Word")}

\item{table_fns_ls}{Table functions (a list), Default: make_table_fns_ls()}

\item{use_lbls_as_col_nms_1L_lgl}{Use labels as column names (a logical vector of length one), Default: T}

\item{var_desc_chr}{Variable description (a character vector), Default: 'NA'}

\item{...}{Additional arguments}
}
\value{
Table (an output object of multiple potential types)
}
\description{
print_in_format() is a Print function that prints output to console. Specifically, this function implements an algorithm to print in format. The function returns Table (an output object of multiple potential types).
}
