% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_plot}
\alias{print_plot}
\title{Print plot}
\usage{
print_plot(
  plot_plt,
  name_1L_chr = "plot",
  output_type_1L_chr = c("PDF", "HTML", "Word"),
  scale_1L_dbl = 1,
  write_to_1L_chr = tempdir()
)
}
\arguments{
\item{plot_plt}{Plot (a plot)}

\item{name_1L_chr}{Name (a character vector of length one), Default: 'plot'}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: c("PDF", "HTML", "Word")}

\item{scale_1L_dbl}{Scale (a double vector of length one), Default: 1}

\item{write_to_1L_chr}{Write to (a character vector of length one), Default: tempdir()}
}
\value{
No return value, called for side effects.
}
\description{
print_plot() is a Print function that prints output to console. Specifically, this function implements an algorithm to print plot. The function is called for its side effects and does not return a value.
}
\keyword{internal}
