% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_authors.R
\name{ready4show_authors}
\alias{ready4show_authors}
\alias{make_pt_ready4show_authors}
\alias{is_ready4show_authors}
\title{ready4 S3 class for authors lookup table}
\usage{
ready4show_authors(x = make_pt_ready4show_authors())

make_pt_ready4show_authors(
  first_nm_chr = character(0),
  middle_nm_chr = character(0),
  last_nm_chr = character(0),
  title_chr = character(0),
  qualifications_chr = character(0),
  institute_chr = character(0),
  sequence_int = integer(0),
  is_corresponding_lgl = logical(0),
  email_chr = character(0),
  is_equal_first_lgl = logical(0)
)

is_ready4show_authors(x)
}
\arguments{
\item{x}{An object of any type}

\item{first_nm_chr}{First name (a character vector), Default: character(0)}

\item{middle_nm_chr}{Middle name (a character vector), Default: character(0)}

\item{last_nm_chr}{Last name (a character vector), Default: character(0)}

\item{title_chr}{Title (a character vector), Default: character(0)}

\item{qualifications_chr}{Qualifications (a character vector), Default: character(0)}

\item{institute_chr}{Institute (a character vector), Default: character(0)}

\item{sequence_int}{Sequence (an integer vector), Default: integer(0)}

\item{is_corresponding_lgl}{Is corresponding (a logical vector), Default: logical(0)}

\item{email_chr}{Email (a character vector), Default: character(0)}

\item{is_equal_first_lgl}{Is equal first (a logical vector), Default: logical(0)}
}
\value{
A validated instance of the ready4 S3 class for authors lookup table

A prototype for ready4 S3 class for authors lookup table

A logical value, TRUE if a valid instance of the ready4 S3 class for authors lookup table
}
\description{
Create a new valid instance of the ready4 S3 class for authors lookup table
}
\details{
ready4 S3 class for authors lookup table
}
