% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_correspondences.R
\name{ready4show_correspondences}
\alias{ready4show_correspondences}
\alias{make_pt_ready4show_correspondences}
\alias{is_ready4show_correspondences}
\title{Name correspondences lookup table}
\usage{
ready4show_correspondences(x = make_pt_ready4show_correspondences())

make_pt_ready4show_correspondences(
  old_nms_chr = character(0),
  new_nms_chr = character(0)
)

is_ready4show_correspondences(x)
}
\arguments{
\item{x}{An object of any type}

\item{old_nms_chr}{Old names (a character vector), Default: character(0)}

\item{new_nms_chr}{New names (a character vector), Default: character(0)}
}
\value{
A validated instance of the Name correspondences lookup table

A prototype for Name correspondences lookup table

A logical value, TRUE if a valid instance of the Name correspondences lookup table
}
\description{
Create a new valid instance of the Name correspondences lookup table
}
\details{
Name correspondences lookup table
}
