% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_institutes.R
\name{ready4show_institutes}
\alias{ready4show_institutes}
\title{ready4 S3 class for institutes lookup table}
\usage{
ready4show_institutes(x = make_pt_ready4show_institutes())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for institutes lookup table, Default: make_pt_ready4show_institutes()}
}
\value{
A validated instance of the ready4 S3 class for institutes lookup table
}
\description{
Create a new valid instance of the ready4 S3 class for institutes lookup table
}
\details{
ready4 S3 class for institutes lookup table
}
