% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4showPaths.R
\docType{class}
\name{Ready4showPaths-class}
\alias{Ready4showPaths-class}
\alias{Ready4showPaths}
\title{Ready4showPaths}
\description{
Metadata about paths to Markdown input and output
}
\section{Slots}{

\describe{
\item{\code{mkdn_data_dir_1L_chr}}{Markdown data directory (a character vector of length one)}

\item{\code{mkdn_source_dir_1L_chr}}{Markdown source directory (a character vector of length one)}

\item{\code{ms_mkdn_dir_1L_chr}}{Manuscript markdown directory (a character vector of length one)}

\item{\code{ms_dir_1L_chr}}{Manuscript directory (a character vector of length one)}

\item{\code{outp_data_dir_1L_chr}}{Output data directory (a character vector of length one)}

\item{\code{reports_dir_1L_chr}}{Reports directory (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

