% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4showSynopsis.R
\docType{class}
\name{Ready4showSynopsis-class}
\alias{Ready4showSynopsis-class}
\alias{Ready4showSynopsis}
\title{Ready4showSynopsis}
\description{
Metadata about a scientific summary manuscript
}
\section{Slots}{

\describe{
\item{\code{authors_r3}}{Authors (a ready4 S3)}

\item{\code{background_1L_chr}}{Background (a character vector of length one)}

\item{\code{coi_1L_chr}}{Conflict of interest (a character vector of length one)}

\item{\code{conclusion_1L_chr}}{Conclusion (a character vector of length one)}

\item{\code{correspondences_r3}}{Correspondences (a ready4 S3)}

\item{\code{digits_int}}{Digits (an integer vector)}

\item{\code{ethics_1L_chr}}{Ethics (a character vector of length one)}

\item{\code{funding_1L_chr}}{Funding (a character vector of length one)}

\item{\code{institutes_r3}}{Institutes (a ready4 S3)}

\item{\code{interval_chr}}{Interval (a character vector)}

\item{\code{keywords_chr}}{Keywords (a character vector)}

\item{\code{outp_formats_chr}}{Output formats (a character vector)}

\item{\code{sample_desc_1L_chr}}{Sample description (a character vector of length one)}

\item{\code{title_1L_chr}}{Title (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

