% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_abstract}
\alias{transform_abstract}
\title{transform abstract}
\usage{
transform_abstract(path_to_abstract_1L_chr)
}
\arguments{
\item{path_to_abstract_1L_chr}{Path to abstract (a character vector of length one)}
}
\value{
Transformed abstract (a character vector)
}
\description{
transform_abstract() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform abstract. Function argument path_to_abstract_1L_chr specifies the object to be updated. The function returns Transformed abstract (a character vector).
}
\keyword{internal}
