% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_validate.R
\name{validate_args_inc_consent}
\alias{validate_args_inc_consent}
\title{Validate arguments include consent}
\usage{
validate_args_inc_consent(
  args_ls,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N")
)
}
\arguments{
\item{args_ls}{Arguments (a list)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\value{
Arguments (a list)
}
\description{
validate_args_inc_consent() is a Validate function that validates that an object conforms to required criteria. Specifically, this function implements an algorithm to validate arguments include consent. The function returns Arguments (a list).
}
\keyword{internal}
