% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_institutes_lup.R
\name{validate_ready4_institutes_lup}
\alias{validate_ready4_institutes_lup}
\title{Validate ready4show S3 class for institutes lookup table}
\usage{
validate_ready4_institutes_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4show S3 class for institutes lookup table}
}
\value{
A prototpe for ready4show S3 class for institutes lookup table
}
\description{
Validate an instance of the ready4show S3 class for institutes lookup table
}
\details{
ready4show S3 class for institutes lookup table
}
