% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_authors_lup.R
\name{validate_ready4show_authors_lup}
\alias{validate_ready4show_authors_lup}
\title{Validate ready4show authors lookup table ready4 S3 class for authors lookup table}
\usage{
validate_ready4show_authors_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for authors lookup table}
}
\value{
A prototpe for ready4 S3 class for authors lookup table
}
\description{
Validate an instance of the ready4 S3 class for authors lookup table
}
\details{
ready4 S3 class for authors lookup table
}
