% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_correspondences.R
\name{validate_ready4show_correspondences}
\alias{validate_ready4show_correspondences}
\title{validate ready4show correspondences Name correspondences lookup table}
\usage{
validate_ready4show_correspondences(x)
}
\arguments{
\item{x}{An unvalidated instance of the Name correspondences lookup table}
}
\value{
A prototpe for Name correspondences lookup table
}
\description{
Validate an instance of the Name correspondences lookup table
}
\details{
Name correspondences lookup table
}
\keyword{internal}
