% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_custom_authors}
\alias{write_custom_authors}
\title{Write custom authors}
\usage{
write_custom_authors(
  paths_ls,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N"),
  rmd_fl_nms_ls = make_rmd_fl_nms_ls()
)
}
\arguments{
\item{paths_ls}{Paths (a list)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{rmd_fl_nms_ls}{R Markdown file names (a list), Default: make_rmd_fl_nms_ls()}
}
\value{
No return value, called for side effects.
}
\description{
write_custom_authors() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write custom authors. The function is called for its side effects and does not return a value.
}
\keyword{internal}
