% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_header_yaml}
\alias{write_header_yaml}
\title{write header yaml}
\usage{
write_header_yaml(
  path_to_header_dir_1L_chr,
  fl_nm_1L_chr = "header.yaml",
  authors_tb,
  institutes_tb,
  title_1L_chr = "Example title",
  keywords_chr = c("Example keyword one", "Example keyword two"),
  path_to_tmpl_header_1L_chr = NULL,
  inc_quals_1L_lgl = F
)
}
\arguments{
\item{path_to_header_dir_1L_chr}{Path to header directory (a character vector of length one)}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'header.yaml'}

\item{authors_tb}{Authors (a tibble)}

\item{institutes_tb}{Institutes (a tibble)}

\item{title_1L_chr}{Title (a character vector of length one), Default: 'Example title'}

\item{keywords_chr}{Keywords (a character vector), Default: c("Example keyword one", "Example keyword two")}

\item{path_to_tmpl_header_1L_chr}{Path to template header (a character vector of length one), Default: NULL}

\item{inc_quals_1L_lgl}{Include qualifications (a logical vector of length one), Default: F}
}
\description{
write_header_yaml() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write header yaml. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
