% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mkdn_from_pkg}
\alias{write_mkdn_from_pkg}
\title{Write markdown from package}
\usage{
write_mkdn_from_pkg(
  pkg_nm_1L_chr,
  destn_dir_1L_chr = "Markdown",
  overwrite_1L_lgl = F
)
}
\arguments{
\item{pkg_nm_1L_chr}{Package name (a character vector of length one)}

\item{destn_dir_1L_chr}{Destn directory (a character vector of length one), Default: 'Markdown'}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: F}
}
\description{
write_mkdn_from_pkg() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write markdown from package. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
