% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_mkdn_from_repo}
\alias{write_mkdn_from_repo}
\title{Write markdown from repository}
\usage{
write_mkdn_from_repo(
  mkdn_data_dir_1L_chr,
  ms_mkdn_parent_1L_chr,
  ms_mkdn_repo_1L_chr,
  version_1L_chr,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  options_chr = c("Y", "N")
)
}
\arguments{
\item{mkdn_data_dir_1L_chr}{Markdown data directory (a character vector of length one)}

\item{ms_mkdn_parent_1L_chr}{Manuscript markdown parent (a character vector of length one)}

\item{ms_mkdn_repo_1L_chr}{Manuscript markdown repository (a character vector of length one)}

\item{version_1L_chr}{Version (a character vector of length one)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}
}
\description{
write_mkdn_from_repo() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write markdown from repository. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
