% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData-Ready4showSynopsis}
\alias{authorData-Ready4showSynopsis}
\alias{authorData,Ready4showSynopsis-method}
\title{Author and document datasets}
\usage{
\S4method{authorData}{Ready4showSynopsis}(
  x,
  tmpl_url_1L_chr = "https://github.com/ready4-dev/ms_tmpl",
  tmpl_version_1_L_chr = "0.0.9.2",
  what_1L_chr = "Manuscript"
)
}
\arguments{
\item{x}{An object of class Ready4showSynopsis}

\item{tmpl_url_1L_chr}{Template url (a character vector of length one), Default: 'https://github.com/ready4-dev/ms_tmpl'}

\item{tmpl_version_1_L_chr}{Template version 1 L (a character vector), Default: '0.0.9.2'}

\item{what_1L_chr}{What (a character vector of length one), Default: 'Manuscript'}
}
\description{
authorData method applied to Ready4showSynopsis
}
