% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorReport.R
\name{authorReport-Ready4showSynopsis}
\alias{authorReport-Ready4showSynopsis}
\alias{authorReport,Ready4showSynopsis-method}
\title{Author and save a report}
\usage{
\S4method{authorReport}{Ready4showSynopsis}(
  x,
  args_ls = NULL,
  consent_1L_chr = "",
  consent_indcs_int = 1L,
  fl_nm_1L_chr = NA_character_,
  options_chr = c("Y", "N"),
  rmd_fl_nms_ls = NULL,
  what_1L_chr = NA_character_,
  ...
)
}
\arguments{
\item{x}{An object of class Ready4showSynopsis}

\item{args_ls}{Arguments (a list), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{consent_indcs_int}{Consent indices (an integer vector), Default: 1}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'NA'}

\item{options_chr}{Options (a character vector), Default: c("Y", "N")}

\item{rmd_fl_nms_ls}{R Markdown file names (a list), Default: NULL}

\item{what_1L_chr}{What (a character vector of length one), Default: 'NA'}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects.
}
\description{
authorReport method applied to Ready4showSynopsis
}
\seealso{
\code{\link[rmarkdown:render]{rmarkdown::render()}}
}
