% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4show_institutes.R
\name{make_new_ready4show_institutes}
\alias{make_new_ready4show_institutes}
\title{Make new ready4show institutes ready4 S3 class for institutes lookup table}
\usage{
make_new_ready4show_institutes(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for institutes lookup table}
}
\value{
An unvalidated instance of the ready4 S3 class for institutes lookup table
}
\description{
Create a new unvalidated instance of the ready4 S3 class for institutes lookup table
}
\details{
ready4 S3 class for institutes lookup table
}
