% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture-Ready4showSynopsis}
\alias{manufacture-Ready4showSynopsis}
\alias{manufacture,Ready4showSynopsis-method}
\title{Manufacture a (non ready4 framework) object}
\usage{
\S4method{manufacture}{Ready4showSynopsis}(x, what_1L_chr = "paths_ls")
}
\arguments{
\item{x}{An object of class Ready4showSynopsis}

\item{what_1L_chr}{What (a character vector of length one), Default: 'paths_ls'}
}
\value{
Object (an output object of multiple potential types)
}
\description{
manufacture method applied to Ready4showSynopsis
}
