% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4show.R
\docType{package}
\name{ready4show-package}
\alias{ready4show}
\alias{ready4show-package}
\title{ready4show: Standardised Developer Tools for Sharing Insights from Projects
Developed with the Ready4 Suite}
\description{
ready4show provides a set of classes and methods for
authoring reports and user-interfaces that showcase insights generated
by projects developped with the ready4 suite of tools for mental
health data synthesis and modelling.  This development version of the
ready4show package has been made available as part of the process of
testing and documenting the package. If you have any questions, please
contact the authors (matthew.hamilton@orygen.org.au).
}
\details{
To learn more about ready4show, start with the vignettes:
\code{browseVignettes(package = "ready4show")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4show/}
  \item \url{https://github.com/ready4-dev/ready4show}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
