% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4show.R, R/ready4show-package.R
\docType{package}
\name{ready4show-package}
\alias{ready4show-package}
\title{ready4show: Author Health Economic Analysis Programs and Reporting Templates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ready4show provides tools for authoring reproducible analyses and to help automate the generation of scientific summaries for modelling studies implemented with the ready4 framework (https://www.ready4-dev.com). This development version of the ready4show package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ready4show provides tools for authoring reproducible analyses and to help automate the generation of scientific summaries for modelling studies implemented with the ready4 framework (https://www.ready4-dev.com). This development version of the ready4show package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4show/}
  \item \url{https://github.com/ready4-dev/ready4show}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4show/}
  \item \url{https://github.com/ready4-dev/ready4show}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Glen Wiesner (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Australian Government Research Training Program [funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
