% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_custom_authors}
\alias{write_custom_authors}
\title{Write custom authors}
\usage{
write_custom_authors(paths_ls)
}
\arguments{
\item{paths_ls}{Paths (a list)}
}
\description{
write_custom_authors() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write custom authors. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
