% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_header_fls}
\alias{write_header_fls}
\title{Write header files}
\usage{
write_header_fls(
  path_to_header_dir_1L_chr,
  header_yaml_args_ls,
  abstract_args_ls = NULL
)
}
\arguments{
\item{path_to_header_dir_1L_chr}{Path to header directory (a character vector of length one)}

\item{header_yaml_args_ls}{Header yaml arguments (a list)}

\item{abstract_args_ls}{Abstract arguments (a list), Default: NULL}
}
\description{
write_header_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write header files. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
