% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_rprt}
\alias{write_rprt}
\title{Write report}
\usage{
write_rprt(
  rprt_type_ls,
  outp_smry_ls,
  output_type_1L_chr = "PDF",
  section_type_1L_chr = "#",
  path_to_prjs_dir_1L_chr = "../../../../Data/Project",
  prt_dir_dir_1L_chr = "My_Project",
  reports_dir_1L_chr = "Reports",
  rltv_path_to_data_dir_1L_chr = "../Output",
  nm_of_mkdn_dir_1L_chr = "Markdown",
  push_copy_to_dv_1L_lgl = T
)
}
\arguments{
\item{rprt_type_ls}{Report type (a list)}

\item{outp_smry_ls}{Output smry (a list)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'PDF'}

\item{section_type_1L_chr}{Section type (a character vector of length one), Default: '#'}

\item{path_to_prjs_dir_1L_chr}{Path to prjs directory (a character vector of length one), Default: '../../../../Data/Project'}

\item{prt_dir_dir_1L_chr}{Prt directory directory (a character vector of length one), Default: 'My_Project'}

\item{reports_dir_1L_chr}{Reports directory (a character vector of length one), Default: 'Reports'}

\item{rltv_path_to_data_dir_1L_chr}{Rltv path to data directory (a character vector of length one), Default: '../Output'}

\item{nm_of_mkdn_dir_1L_chr}{Name of markdown directory (a character vector of length one), Default: 'Markdown'}

\item{push_copy_to_dv_1L_lgl}{Push copy to dataverse (a logical vector of length one), Default: T}
}
\value{
Output smry (a list)
}
\description{
write_rprt() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write report. The function returns Output smry (a list).
}
\keyword{internal}
